Pi/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1
pyserial==3.5
